requirements.txt
# Core dependencies
Flask==2.3.3
flask-cors==4.0.2
flask-swagger-ui==4.11.1
psycopg2-binary>=2.9.9
python-dotenv==1.0.0
gunicorn>=22.0.1
requests==2.32.3
prometheus-client==0.16.0
urllib3>=2.0.7
Jinja2==3.1.6
Werkzeug==3.0.6
MarkupSafe==3.0.2

# Development and testing dependencies
pytest>=7.4.0
pytest-cov==4.1.0
pytest-env==0.8.1
black==24.3.0
flake8==6.0.0
isort==5.12.0
termcolor==2.3.0
pyyaml==6.0.1

# Automation and monitoring dependencies
aiohttp==3.9.3
selenium==4.27.1
playwright==1.40.0

# Documentation
mkdocs==1.6.1
mkdocs-material==9.6.12
mkdocs-minify-plugin==0.7.1
